orpsocv2: correct build/par issue on Atlys board
From patch submission e-mail:

The first of the series is an issue that has been around since the
beginning (at least based on the svn log files) and traces its root to the
first Xilinx board supported by orpsocv2, the ML501. Apparently Makefile
for the final place-and-route process in all the builds contains a typo
which leads to the routing tool not using any timing constraint at all.

Patch by: Jason Zheng <jxzheng@gmail.com>

git-svn-id: http://opencores.org/ocsvn/openrisc/openrisc@866 04c7f051-eaf3-4997-922d-f29d9d8c0f3f